fixed number casting problem

+++ impl/src/main/java/org/dukecon/server/eventbooking/EventBookingResource.groovy
@@ -44,7 +44,7 @@ class EventBookingResource {
     Response getBookingInformation(@PathParam("conferenceId") String conferenceId) {
         def allCapacities = bookingService.getAllCapacities(conferenceId)
         def events = conferences[conferenceId].events.collect { Event e ->
-            allCapacities*.eventId.contains(e.id) ? allCapacities.find {e.id == it.eventId}.withNumberOfFavoritesAndLocationCapacity(preferencesService.allEventFavorites[e.id] ?: 0, e.location?.capacity ?: 0) : new EventBooking(conferenceId: conferenceId, eventId: e.id, locationCapacity: e.location?.capacity, numberOfFavorites: preferencesService.allEventFavorites[e.id] ?: 0)
+            allCapacities*.eventId.contains(e.id) ? allCapacities.find {e.id == it.eventId}.withNumberOfFavoritesAndLocationCapacity((preferencesService.allEventFavorites[e.id] ?: 0) as int, e.location?.capacity ?: 0) : new EventBooking(conferenceId: conferenceId, eventId: e.id, locationCapacity: e.location?.capacity, numberOfFavorites: preferencesService.allEventFavorites[e.id] ?: 0)
         }
         return Response.ok(events).build()
     }
